requirements.txt
anyio==4.12.1
APScheduler==3.11.2
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
python-dotenv==1.2.1
python-http-client==3.3.7
python-telegram-bot==22.6
requests==2.32.5
resend==2.23.0
sendgrid==6.12.5
SQLAlchemy==2.0.47
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.6
